+++ pkg/cniplugin/cni.go
@@ -16,18 +16,18 @@ import (
 )
 
 type CniOpts struct {
-	WaitForUdev       bool
-	WaitForUdevPrefix string
-	WaitForUdevDelay  time.Duration
-	WaitForUdevTimeout  time.Duration
+	WaitForUdev        bool
+	WaitForUdevPrefix  string
+	WaitForUdevDelay   time.Duration
+	WaitForUdevTimeout time.Duration
 }
 
 func DefaultCniOpts() CniOpts {
 	return CniOpts{
-		WaitForUdev:       true,
-		WaitForUdevPrefix: "eth",
-		WaitForUdevDelay:  100 * time.Millisecond,
-		WaitForUdevTimeout:  5000 * time.Millisecond,
+		WaitForUdev:        true,
+		WaitForUdevPrefix:  "eth",
+		WaitForUdevDelay:   100 * time.Millisecond,
+		WaitForUdevTimeout: 5000 * time.Millisecond,
 	}
 }