+++ pkg/cniserver/cmdhandler_test.go
@@ -34,7 +34,7 @@ func Test_CmdHandler(t *testing.T) {
 			Assert(t).That(results, Not(IsNil()))
 
 			// ensure the port exists
-			port, err := deps.OpenstackClient().GetPortByTags(cniserver.NewPortTags(cmd).AsStringSlice())
+			port, err := deps.OpenstackClient().GetPortByTags(cniserver.NewPortTagsFromCommand(cmd).NeutronTags().AsStringSlice())
 			Assert(t).That(err, IsNil())
 			Assert(t).That(port, Not(IsNil()))
 
@@ -42,7 +42,7 @@ func Test_CmdHandler(t *testing.T) {
 			Assert(t).That(deps.CniHandler().Del(cmd), IsNil())
 
 			// ensure the port's gone
-			port, perr := deps.OpenstackClient().GetPortByTags(cniserver.NewPortTags(cmd).AsStringSlice())
+			port, perr := deps.OpenstackClient().GetPortByTags(cniserver.NewPortTagsFromCommand(cmd).NeutronTags().AsStringSlice())
 			Assert(t).That(perr, Equals(openstack.ErrPortNotFound))
 		})
 	})
@@ -60,14 +60,14 @@ func Test_NewPortTags(t *testing.T) {
 	t.Run("ensure container ids are shortened to the first 12 chars", func(t *testing.T) {
 		cmd := NewTestData().CniCommand()
 		cmd.ContainerID = "12345678901234567890"
-		tags := cniserver.NewPortTags(cmd)
+		tags := cniserver.NewPortTagsFromCommand(cmd).NeutronTags()
 		Assert(t).That(tags.Tags[0], Equals("containerid=123456789012"))
 	})
 
 	t.Run("ensure a short container id doesn't fail.. this shouldn't be possible in k8s but we'll play it safe", func(t *testing.T) {
 		cmd := NewTestData().CniCommand()
 		cmd.ContainerID = "123"
-		tags := cniserver.NewPortTags(cmd)
+		tags := cniserver.NewPortTagsFromCommand(cmd).NeutronTags()
 		Assert(t).That(tags.Tags[0], Equals("containerid=123"))
 	})
 }