+++ pkg/openstack/port_manager_test.go
@@ -68,7 +68,7 @@ func SetupAndTeardownPort(t *testing.T, context util.CniContext, client openstac
 	t.Helper()
 	pm := openstack.NewPortManager(client)
 	opts := openstack.SetupPortOptsFromContext(context)
-	opts.Tags = cniserver.NewPortTags(context.Command)
+	opts.Tags = cniserver.NewPortTagsFromCommand(context.Command).NeutronTags()
 
 	results, err := pm.SetupPort(opts)
 	Assert(t).That(err, IsNil(), "failed to setup port")
@@ -81,7 +81,7 @@ func SetupAndTeardownPort(t *testing.T, context util.CniContext, client openstac
 	_, err = client.GetPortByTags(opts.Tags.AsStringSlice())
 	Assert(t).That(err, IsNil(), "failed get port by tags %s", opts.Tags.String())
 
-	tdOpts := openstack.TearDownPortOpts{Hostname: context.Hostname, Tags: cniserver.NewPortTags(context.Command)}
+	tdOpts := openstack.TearDownPortOpts{Hostname: context.Hostname, Tags: cniserver.NewPortTagsFromCommand(context.Command).NeutronTags()}
 	err = pm.TeardownPort(tdOpts)
 	Assert(t).That(err, IsNil(), "failed teardown port")
 
@@ -96,7 +96,7 @@ func SetupAndTeardownPort(t *testing.T, context util.CniContext, client openstac
 	_, err = client.GetPortByTags(opts.Tags.AsStringSlice())
 	Assert(t).That(err, IsNil(), "failed get port by tags %s", opts.Tags.String())
 
-	tdOpts = openstack.TearDownPortOpts{Hostname: context.Hostname, Tags: cniserver.NewPortTags(context.Command)}
+	tdOpts = openstack.TearDownPortOpts{Hostname: context.Hostname, Tags: cniserver.NewPortTagsFromCommand(context.Command).NeutronTags()}
 	err = pm.TeardownPort(tdOpts)
 	Assert(t).That(err, IsNil(), "failed teardown port")