+++ pkg/util/util.go
@@ -2,8 +2,10 @@ package util
 
 import (
 	"encoding/json"
+	"fmt"
 	"os"
 	"strconv"
+	"time"
 )
 
 func FromJson(data []byte, i any) error {
@@ -34,7 +36,20 @@ func GetenvAsBool(key string, def bool) bool {
 	return r
 }
 
-func fileEntryExists(path string, isDir bool) (bool, error) {
+func GetenvAsDuration(key string, def time.Duration) time.Duration {
+	v := os.Getenv(key)
+	if v == "" {
+		return def
+	}
+
+	r, err := time.ParseDuration(fmt.Sprintf("%sms", v))
+	if err != nil {
+		return def
+	}
+	return r
+}
+
+func FileExists(path string) (bool, error) {
 	info, err := os.Stat(path)
 	if err != nil {
 		if os.IsNotExist(err) {
@@ -45,17 +60,20 @@ func fileEntryExists(path string, isDir bool) (bool, error) {
 		}
 		return false, err
 	}
+	isDir := info.IsDir()
+	return !isDir, nil
+}
 
-	if isDir {
-		return info.IsDir(), nil
+func DirExists(path string) (bool, error) {
+	info, err := os.Stat(path)
+	if err != nil {
+		if os.IsNotExist(err) {
+			return false, nil
+		}
+		if os.IsPermission(err) {
+			return true, nil
+		}
+		return false, err
 	}
-	return !info.IsDir(), nil
-}
-
-func FileExists(file string) (bool, error) {
-	return fileEntryExists(file, false)
-}
-
-func DirExists(dir string) (bool, error) {
-	return fileEntryExists(dir, true)
+	return info.IsDir(), nil
 }